pyproject.toml
[project]
name = "qracer"
version = "0.1.0"
description = "Conversational investment agent for discovering alpha in global security markets"
requires-python = ">=3.12"
dependencies = [
  "click>=8.1.0",
  "yfinance>=0.2.0",
  "duckdb>=1.2.0",
  "pydantic>=2.0.0",
  "python-dotenv>=1.0.0",
  "tomli>=2.0.0; python_version < '3.11'",
]

[dependency-groups]
dev = [
  "ruff>=0.9.0",
  "pyright>=1.1.0",
  "pytest>=8.0.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
]

[project.scripts]
qracer = "qracer.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingTypeStubs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
